requirements.txt
ollama>=0.1.0
google-generativeai>=0.3.0
mistralai>=0.0.8
requests>=2.31.0
pydantic>=2.5.0
pyyaml>=6.0
reportlab>=4.0.0
markdown>=3.5.0
python-dotenv>=1.0.0
rich>=13.7.0
